+++ CHANGELOG.rst
@@ -1,3 +1,10 @@
+HEAD (2026-??-??)
+-----------------
+
+*	Added the ``Control`` attributes ``in_search``, ``in_fulltext_search`` and
+	``in_expert_search``.
+
+
 0.53.1 (2025-12-15)
 -------------------